Views & CSS/Images
Now that we've gotten the Gemfile taken care of so DotCloud can install
our dependencies we add the last parts of our application. We're adding
all the views used, as well as the images and CSS on the page.

This tutorial is using Haml as the markup language but you could
substitute your preferred language (erb, slim, etc). Let's talk just a
bit about some of the views.

First, ``layout.haml`` serves to generate the "layout" of our site or
"what goes where" if you will. Thanks to Haml we save quite a bit of
typing. Sinatra by default looks for a file named layout in the
``views/`` directory to serve as the applications layout, we don't need
to do anything special to get this to work.

The ``index.haml`` gives us the header text, image, and 'view source'
link and tells Sinatra to render our two remaining templates, text and
footer.

``text.haml`` is the form that contains the input fields for "Number"
and "Message" as well as the "Text Me!" button. Finally ``footer.haml``
is made up of just the Twitter logo. It might be overkill in this simple
example to give this its own template but if the application were to
grow it makes it a little easier to add things and keeps our templates small.

With these files added you can now ``dotcloud push cloudsinger`` and
then visit the URL of your deployed app. When you visit the URL you
should be able to supply a phone number and a message (140 characters or
less of course!) that you would like to send. Click on "Text Me!" and if
you've setup ``config.yml`` correctly and all systems are go, you should
get a text shortly at the number you provided. Please note that if
you're using a Twilio demo account you may need to edit ``app.rb`` line
33. In this file you will need to add your Twilio PIN as the first
argument to ``Twilio::Sms.message()``, this is also documented directly
in ``app.rb``.

This is just the beginning of what's possible with Twilio. You can make
and receive phone calls, record them, create conferences, and receive
SMS messages. In a non-trivial application you would login to the Twilio
dashboard and setup Twilio to HTTP POST to a URL of your app when
someone calls or SMS messages your application and you would then write
the code to handle those POST events.

I hope you've enjoyed this DotCloud + Twilio demo, feel free to drop us
a line if you have problems.